Separate seasonal risk to separate component
